fix options passing in pipeline (#18)

+++ src/main/java/com/google/exposurenotification/privateanalytics/ingestion/FirestoreConnector.java
@@ -122,16 +122,15 @@ public class FirestoreConnector {
 
       @ProcessElement
       public void processElement(ProcessContext context) {
-        IngestionPipelineOptions options = context.getPipelineOptions()
-            .as(IngestionPipelineOptions.class);
+        IngestionPipelineOptions options = context.getPipelineOptions().as(IngestionPipelineOptions.class);
         long startTime = options.getStartTime().get();
-        int backwardWindow = (int) (options.getGracePeriodBackwards().get() / SECONDS_IN_HOUR);
-        int forwardWindow = (int) (options.getGracePeriodForwards().get() / SECONDS_IN_HOUR);
+        long backwardWindow = options.getGracePeriodBackwards().get() / SECONDS_IN_HOUR;
+        long forwardWindow = options.getGracePeriodForwards().get() / SECONDS_IN_HOUR;
 
         // Each datashare in Firestore is stored under a Date collection with the format: yyyy-MM-dd-HH.
         // To query all documents uploaded around startTime within the specified window, construct
         // a query for each hour within the window: [startTime - backwardWindow, startTime + forwardWindow].
-        for (int i = (-1 * backwardWindow); i <= forwardWindow; i++) {
+        for (long i = (-1 * backwardWindow); i <= forwardWindow; i++) {
           long timeToQuery = startTime + i * SECONDS_IN_HOUR;
           LocalDateTime dateTimeToQuery = LocalDateTime
               .ofEpochSecond(timeToQuery, 0, ZoneOffset.UTC);
